+++ crates/vm/levm/src/opcode_handlers/stack_memory_storage_flow.rs
@@ -23,7 +23,7 @@ impl<'a> VM<'a> {
     pub fn op_pop(&mut self) -> Result<OpcodeResult, VMError> {
         let current_call_frame = self.current_call_frame_mut()?;
         current_call_frame.increase_consumed_gas(gas_cost::POP)?;
-        current_call_frame.stack.pop::<1>()?;
+        current_call_frame.stack.pop1()?;
         Ok(OpcodeResult::Continue { pc_increment: 1 })
     }
 
@@ -34,7 +34,7 @@ impl<'a> VM<'a> {
             return Err(ExceptionalHalt::InvalidOpcode.into());
         }
 
-        let [key] = *self.current_call_frame_mut()?.stack.pop()?;
+        let key = self.current_call_frame_mut()?.stack.pop1()?;
         let to = self.current_call_frame()?.to;
         let value = self
             .substate
@@ -47,7 +47,7 @@ impl<'a> VM<'a> {
 
         current_call_frame.increase_consumed_gas(gas_cost::TLOAD)?;
 
-        current_call_frame.stack.push(&[value])?;
+        current_call_frame.stack.push1(value)?;
         Ok(OpcodeResult::Continue { pc_increment: 1 })
     }
 
@@ -77,7 +77,7 @@ impl<'a> VM<'a> {
     // MLOAD operation
     pub fn op_mload(&mut self) -> Result<OpcodeResult, VMError> {
         let current_call_frame = self.current_call_frame_mut()?;
-        let [offset] = *current_call_frame.stack.pop()?;
+        let offset = current_call_frame.stack.pop1()?;
 
         let offset: usize = offset
             .try_into()
@@ -92,7 +92,7 @@ impl<'a> VM<'a> {
 
         current_call_frame
             .stack
-            .push(&[current_call_frame.memory.load_word(offset)?])?;
+            .push1(current_call_frame.memory.load_word(offset)?)?;
 
         Ok(OpcodeResult::Continue { pc_increment: 1 })
     }
@@ -128,7 +128,7 @@ impl<'a> VM<'a> {
     pub fn op_mstore8(&mut self) -> Result<OpcodeResult, VMError> {
         let current_call_frame = self.current_call_frame_mut()?;
 
-        let [offset] = *current_call_frame.stack.pop()?;
+        let offset = current_call_frame.stack.pop1()?;
 
         let offset: usize = offset
             .try_into()
@@ -141,12 +141,11 @@ impl<'a> VM<'a> {
             current_call_frame.memory.len(),
         )?)?;
 
-        let [value] = current_call_frame.stack.pop()?;
+        let value = current_call_frame.stack.pop1()?;
 
-        current_call_frame.memory.store_data(
-            offset,
-            &u256_to_big_endian(*value)[WORD_SIZE - 1..WORD_SIZE],
-        )?;
+        current_call_frame
+            .memory
+            .store_data(offset, &u256_to_big_endian(value)[WORD_SIZE - 1..WORD_SIZE])?;
 
         Ok(OpcodeResult::Continue { pc_increment: 1 })
     }
@@ -155,7 +154,7 @@ impl<'a> VM<'a> {
     pub fn op_sload(&mut self) -> Result<OpcodeResult, VMError> {
         let (storage_slot_key, address) = {
             let current_call_frame = self.current_call_frame_mut()?;
-            let [storage_slot_key] = *current_call_frame.stack.pop()?;
+            let storage_slot_key = current_call_frame.stack.pop1()?;
             let address = current_call_frame.to;
             (storage_slot_key, address)
         };
@@ -168,7 +167,7 @@ impl<'a> VM<'a> {
 
         current_call_frame.increase_consumed_gas(gas_cost::sload(storage_slot_was_cold)?)?;
 
-        current_call_frame.stack.push(&[value])?;
+        current_call_frame.stack.push1(value)?;
         Ok(OpcodeResult::Continue { pc_increment: 1 })
     }
 
@@ -259,7 +258,7 @@ impl<'a> VM<'a> {
         current_call_frame.increase_consumed_gas(gas_cost::MSIZE)?;
         current_call_frame
             .stack
-            .push(&[current_call_frame.memory.len().into()])?;
+            .push1(current_call_frame.memory.len().into())?;
         Ok(OpcodeResult::Continue { pc_increment: 1 })
     }
 
@@ -270,7 +269,7 @@ impl<'a> VM<'a> {
 
         let remaining_gas = current_call_frame.gas_remaining;
         // Note: These are not consumed gas calculations, but are related, so I used this wrapping here
-        current_call_frame.stack.push(&[remaining_gas.into()])?;
+        current_call_frame.stack.push1(remaining_gas.into())?;
 
         Ok(OpcodeResult::Continue { pc_increment: 1 })
     }
@@ -322,7 +321,7 @@ impl<'a> VM<'a> {
         let current_call_frame = self.current_call_frame_mut()?;
         current_call_frame.increase_consumed_gas(gas_cost::JUMP)?;
 
-        let [jump_address] = *current_call_frame.stack.pop()?;
+        let jump_address = current_call_frame.stack.pop1()?;
         Self::jump(current_call_frame, jump_address)?;
 
         Ok(OpcodeResult::Continue { pc_increment: 0 })
@@ -393,7 +392,7 @@ impl<'a> VM<'a> {
 
         current_call_frame
             .stack
-            .push(&[U256::from(current_call_frame.pc)])?;
+            .push1(U256::from(current_call_frame.pc))?;
 
         Ok(OpcodeResult::Continue { pc_increment: 1 })
     }